requirements.txt
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-jsonfield==1.0.1
docopt==0.6.2
facebook-sdk==3.1.0
geographiclib==1.49
geopy==1.18.1
idna==2.8
jsonfield==2.0.2
oauthlib==3.0.1
Pillow==5.4.1
Pillow-PIL==0.1.dev0
pipreqs==0.4.9
PyJWT==1.7.1
pysolr==3.8.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
soupsieve==1.8
urllib3==1.24.1
webpreview==1.5.0
yarg==0.1.9
mysqlclient==1.4.0

